+++ services/__init__.py
@@ -1,0 +1,6 @@
+# Services package initialization
+from .database import DatabaseService
+from .ai_service import AIService, DeepSeekClient
+from .business_service import BusinessService
+
+__all__ = ['DatabaseService', 'AIService', 'DeepSeekClient', 'BusinessService']
